+++ CHANGELOG.md
@@ -6,6 +6,10 @@ DISCLAIMER: Reworked skill ID, and node IDs!
   - Spell modifiers: `class_tier_{tier}_spell_{choice}_modifier_{modifier}` - for example `archer_tier_3_spell_1_modifier_2`
   - Passive spells: `class_tier_{tier}_passive_{choice}` - for example `archer_tier_3_passive_1`
   - Also applies to skill tree node IDs
+- Reworked skill tree tab ID: 
+  - `skill_tree_rpgs:skill_tree_rpgs` -> `skill_tree_rpgs:class_skills`
+- Added new skill tree tab to host independent weapon skills:
+  - ID: `skill_tree_rpgs:weapon_skills`
 - Fixed missing cooldowns on some warrior passives
 - Add conditional attributes feature